pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["AlloKuz <a.o.kuzmin.rus@gmail.com>"]
readme = "README.md"
packages = [{ include = "gendiff" }]
repository = "https://github.com/remortalite/python-project-50"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest-cov = "^5.0.0"
pytest = "^8.2.2"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
